Refactored for future QPC-C introduction
